+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 
 - **Samplesheets**
     - ``template_output_dir_display`` user setting (#1960)
+    - Display BAM/CRAM/VCF omit patterns in study shortcut modal (#1963)
 - **Taskflowbackend**
     - ``BatchCalculateChecksumTask`` retrying in case of timeouts (#1941)
 
@@ -31,6 +32,7 @@ Changed
     - Sanitize iRODS paths in ``get_row_path()`` calls (#1947)
     - ``index`` arg in ``SampleSheetAssayPluginPoint.update_row()`` (#1957)
     - Hide template output dir field by default (#1960)
+    - Improve ``StudyLinksAjaxView`` return data (#1963, #1966)
 - **Taskflowbackend**
     - Increase default for ``TASKFLOW_IRODS_CONN_TIMEOUT`` (#1900)
     - Disable lock requirement for role and project update flows (#1948)
@@ -45,6 +47,7 @@ Fixed
     - Zone with ``NOT CREATED`` status displayed as active in project list (#1962)
 - **Samplesheets**
     - Invalid assay measurement type in ``i_minimal*`` test data (#1954)
+    - Error message handling in ``StudyShortcutModal`` (#1965)
 - **Taskflowbackend**
     - Malformed exception message in ``BatchValidateChecksumsTask`` (#1943)
     - Exceeded zone status info char limit in ``_raise_flow_exception()`` (#1953)